+++ src/plugins/use.ts
@@ -1,6 +1,6 @@
 import Model from '../Model'
-import Repo from '../Repo'
-import Query from '../Query'
+import Repo from '../repo/Repo'
+import Query from '../repo/Query'
 
 export interface Components {
   Model: typeof Model
