+++ src/repo/Incrementer.ts
@@ -1,7 +1,7 @@
 import * as _ from '../support/lodash'
 import { Records } from '../Data'
 import { Increment } from '../Attributes/Types'
-import Repo from '../Repo'
+import Repo from './Repo'
 
 export default class Incrementer {
   /**
